requirements.txt
async-timeout==5.0.1
bleak==2.1.0
dbus-fast==3.1.2
paho-mqtt==2.1.0
typing_extensions==4.15.0
